0-memset.c

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -13,7 +13,7 @@ char *_memset(char *s, char b, unsigned int n)
 	int i;
 	char *ptr = s;
 
-	for (i = 0; i < n; i++)
+	for (i = 0; i < (int) n; i++)
 	{
 		ptr[i] = b;
 	}
